Splitting groupnorm and instancenorm.
